Add Directory for Figures for Slowing Chapter
This is essential for keeping order in the main directory and avoiding naming confusion. It may be necessary to move the tex file for the chapter into this directory as well, which could also facilitate having a way to compile chapters individually, which would also allow tex-pdf syncing and may be critical when the thesis is longer.